requirements.txt
# Flask及其扩展
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Mail==0.9.1
Flask-Bcrypt==1.0.1
email-validator==2.1.0

# 数据库
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
pymysql==1.1.0

# 工具
python-dotenv==1.0.0
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
gunicorn==21.2.0

# 日期和时间处理
pytz==2023.3

# 表单和验证
WTForms==3.1.1

# 其他
Pillow==10.1.0
python-slugify==8.0.1
Werkzeug==2.3.7
SQLAlchemy==2.0.23
alembic==1.12.1
WTForms==3.1.1
requests==2.31.0
gunicorn==21.2.0